Crate oe_rpc_servers

OpenEthereum RPC Servers (WS, HTTP, IPC).

Structs

MetaIoHandler

Request handler

Traits

Metadata

Metadata trait

Middleware

RPC middleware

Functions

start_http

Start http server asynchronously and returns result with Server handle on success or an error.

start_http_with_middleware

Same as start_http, but takes an additional middleware parameter that is introduced as a hyper middleware.

start_ipc

Start IPC server listening on given path.

start_ws

Start WS server and return Server handle.

Type Definitions

HttpServer

Type alias for http server

IpcServer

Type alias for ipc server

WsServer

Type alias for ws server